pyproject.toml
[tool.poetry]
name = "rag-project"
version = "0.1.0"
description = ""
authors = ["Silas Furu Friby <silasfuru.friby@thomsonreuters.com>"]
readme = "README.md"
packages = [{include = "rag_project"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
transformers = "^4.36.0"
datasets = "^2.15.0"
faiss-cpu = "^1.7.4"
torch = "^2.1.1"
openai = "^1.3.8"
python-dotenv = "^1.0.0"
sentence-transformers = "^2.2.2"
llama-index = "0.9.29"
streamlit = "^1.29.0"
playwright = "^1.40.0"
langchain = "^0.0.350"
beautifulsoup4 = "^4.12.2"
qdrant-client = "^1.7.0"
accelerate = "^0.25.0"
huggingface-hub = "^0.20.1"
ipython = "^8.18.1"
requests = "^2.31.0"
guidance = "^0.1.10"
llama-hub = "^0.0.65"
matplotlib = "^3.8.2"
rank-bm25 = "^0.2.2"
huggingface-cli = "^0.1"
logging = "^0.4.9.6"
pinecone-client = "^2.2.4"
chromadb = "^0.4.22"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
